requirements.txt
Flask==2.2.2
flask-apidoc==1.3.0
Flask-Login==0.6.2
flask-restplus==0.13.0
flask-swagger==0.2.14
Flask-WTF==1.1.1
jsonify==0.5
PyJWT==2.6.0
SQLAlchemy==1.4.46
Werkzeug==2.2.2
gunicorn==20.1.0
